color: downgrade "always" to "auto" only for on-disk configuration
An earlier patch downgraded "always" that comes via the ui.color
configuration variable to "auto", in order to work around an
unfortunate regression to "git add -p" that comes from two earlier
mistakes that we no longer undo.  That change however regressed
other third-party tools that depend on "git -c color.ui=always cmd"
as the way to defeat any end-user configuration and force coloured
output from git subcommand, even when the output does not go to a
terminal.

Treating "-c color.ui=always" from the command line differently from
a definition that comes from on-disk configuration files is a bit
ugly, but it is a moral equivalent of "--color=always" option given
to the subcommand from the command line, i.e. a signal that tells us
that the script writer knows what s/he is doing.  So let's take that
route to unbreak this case while defeating a (now declared to be)
misguided color.ui that is set to always in the configuration file.

Signed-off-by: Junio C Hamano <gitster@pobox.com>